[CHORE] 모달 수정 (#119)
