+++ FoodBowl/Global/Base/BaseViewController.swift
@@ -22,6 +22,43 @@ class BaseViewController: UIViewController {
 
     private var activeTextField: UITextField?
 
+    lazy var plusButton = PlusButton().then {
+        let action = UIAction { [weak self] _ in
+            let addFeedViewController = NewFeedViewController()
+            let navigationController = UINavigationController(rootViewController: addFeedViewController)
+            navigationController.modalPresentationStyle = .fullScreen
+            DispatchQueue.main.async {
+                self?.present(navigationController, animated: true)
+            }
+        }
+        $0.addAction(action, for: .touchUpInside)
+    }
+
+    lazy var settingButton = SettingButton().then {
+        let action = UIAction { [weak self] _ in
+            let settingViewController = SettingViewController()
+            self?.navigationController?.pushViewController(settingViewController, animated: true)
+        }
+        $0.addAction(action, for: .touchUpInside)
+    }
+
+    lazy var optionButton = OptionButton().then {
+        let optionButtonAction = UIAction { [weak self] _ in
+            let alert = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+
+            let report = UIAlertAction(title: "신고하기", style: .destructive, handler: { _ in
+                self?.sendReportMail()
+            })
+            let cancel = UIAlertAction(title: "취소", style: .cancel, handler: nil)
+
+            alert.addAction(cancel)
+            alert.addAction(report)
+
+            self?.present(alert, animated: true, completion: nil)
+        }
+        $0.addAction(optionButtonAction, for: .touchUpInside)
+    }
+
     // MARK: - life cycle
     override func viewDidLoad() {
         super.viewDidLoad()
